i45: Fix firebase auth duplication and move initial state setup up to index.js
